requirements.txt
transformers
fastapi
uvicorn
qdrant-client
pillow
torch
torchvision